Mejora del contol y velocidad de obturacion

+++ Sources/APP/Coeficientes.h
@@ -7,10 +7,10 @@
 #ifndef COEFICIENTES_H_
 #define COEFICIENTES_H_
 
-extern short servomotor[1001];
+extern float servomotor[1001];
 extern short motor_izquierdo[1001];
 extern short motor_derecho[1001];
 
-extern short ServoValorMax;
+extern float ServoValorMax;
 
 #endif
